+++ packages/core/command-local-sidecar/src/index.ts
@@ -4,20 +4,18 @@
     - maintains opc connections
     - provides access for native frontend
     - registers subscriptions
-*/  
+*/
 import EventEmitter from 'events'
 import express from 'express';
 import bodyParser from 'body-parser'
 import cors from 'cors';
-import {Server} from 'socket.io'
+import { Server } from 'socket.io'
 
 import { Sidecar } from './wrapper';
 
 const OPC_PROXY_PORT = 8484;
 
-const sidecar = new Sidecar({
-
-});
+const sidecar = new Sidecar();
 
 const app = express();
 
@@ -29,7 +27,7 @@ const io = new Server(server, {
         fn(null, true)
     },
     cors: {
-        
+
     }
 });
 
@@ -44,9 +42,9 @@ io.on('connection', (socket) => {
 //     sidecar.publish_data(data.key, data.value);
 // });
 
-let subscriptions : {
-    events: EventEmitter, 
-    paths: {tag: string, path: string}[], 
+let subscriptions: {
+    events: EventEmitter,
+    paths: { tag: string, path: string }[],
     unsubscribe?: () => void
 } | undefined = undefined;
 
@@ -60,11 +58,26 @@ const dataChanged = (data: any) => {
 app.use(bodyParser.json());
 app.use(cors());
 
-app.route('/setup')
+app.route('/healthcheck')
     .get((req, res) => {
-        const config = sidecar.getConfig();
+        res.send({running: true});
+    })
 
-        res.send(config ? {config} : {error: "No config"}) 
+app.route('/setup')
+    .get(async (req, res) => {
+        const mainConfig = sidecar.getConfig();
+        const config = mainConfig?.iot;
+
+        if (mainConfig && mainConfig.opcuaServer && config) {
+
+            await sidecar.updateConfig(mainConfig);
+
+            await sidecar.connect(mainConfig.opcuaServer)
+
+            await sidecar.setup_data(config.host, config.user, config.pass, config.exchange)
+        }
+
+        res.send(config ? { config } : { error: "No config" })
     })
     .post(async (req, res) => {
         const mainConfig = req.body.config;
@@ -73,11 +86,14 @@ app.route('/setup')
 
         const config = mainConfig.iot;
 
+        sidecar.stop()
+
         await sidecar.connect(mainConfig.opcuaServer)
 
         await sidecar.setup_data(config.host, config.user, config.pass, config.exchange);
 
-        res.send({config: sidecar.getConfig()})
+        await sidecar.start()
+        res.send({ config: sidecar.getConfig() })
     })
 
 app.route('/:host/set_data')
@@ -91,14 +107,14 @@ app.route('/:host/set_data')
         // const client = await sidecar.connect(req.params.host)
 
         // const { type: dt, isArray } = await sidecar.getDataType(client, path)
-        
+
         // if(!dt) return res.send({error: "No datatype"})
 
         // // const client = await sidecar.connect(req.params.host)
         // //TODO pickup dataType from somewhere dynamic
         // const code = await sidecar.setData(client, path, (DataType as any)[dt as any], value);
 
-        res.send({code})
+        res.send({ code })
     })
 
 app.post('/:host/subscribe', async (req, res) => {
@@ -110,12 +126,12 @@ app.post('/:host/subscribe', async (req, res) => {
 
     //     eventEmitter.removeListener('data-changed', dataChanged);
     //     unsubscribe() 
-        
+
     //     subscriptions = undefined
     //     // delete subscriptions
     // }
 
-    if(subscriptions) {
+    if (subscriptions) {
         // current_data[]
         return res.send({
             data: sidecar.values
@@ -124,51 +140,51 @@ app.post('/:host/subscribe', async (req, res) => {
 
     const { subscriptionMap, deviceMap } = sidecar.getConfig() || {}
 
-        try{
-            const client = await sidecar.connect(req.params.host)
+    try {
+        const client = await sidecar.connect(req.params.host)
 
-            const {emitter: events, unsubscribe} = await sidecar.subscribe(subscriptionMap || [])
+        const { unsubscribe } = await sidecar.subscribe(subscriptionMap || [])
 
-            console.log("Subscribed to", subscriptionMap)
+        console.log("Subscribed to", subscriptionMap)
 
-            events.on('data-changed', dataChanged)
+        sidecar.on('data-changed', dataChanged)
 
-            subscriptions = {
-                events,
-                paths: subscriptionMap || [],
-                unsubscribe
-            };
-        }catch(e: any){
-            return res.send({error: e.message})
-        }
+        subscriptions = {
+            events: sidecar,
+            paths: subscriptionMap || [],
+            unsubscribe
+        };
+    } catch (e: any) {
+        return res.send({ error: e.message })
+    }
 
-    
 
-    res.send({success: true})
+
+    res.send({ success: true })
 })
 
 app.post('/:host/unsubscribe', async (req, res) => {
-    if(!subscriptions) return res.send({error: "No subscription found"});
+    if (!subscriptions) return res.send({ error: "No subscription found" });
 
-    try{
-        const {events: eventEmitter, unsubscribe} = subscriptions;
+    try {
+        const { events: eventEmitter, unsubscribe } = subscriptions;
 
         eventEmitter.removeListener('data-changed', dataChanged);
         unsubscribe?.()
-    }catch(e: any){
-        return res.send({error: e.message})
+    } catch (e: any) {
+        return res.send({ error: e.message })
     }
-    res.send({success: true})
+    res.send({ success: true })
 })
 
 app.get('/:host/tree', async (req, res) => {
-    try{
+    try {
         const client = await sidecar.connect(req.params.host);
 
         const tree = await sidecar.browse(client, '/Objects', true, true);
-        res.send({results: tree})
-    }catch(e: any){
-        return res.send({error: e.message})
+        res.send({ results: tree })
+    } catch (e: any) {
+        return res.send({ error: e.message })
     }
 });