+++ packages/communication/opc-amqp-bridge/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.208](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.207...v1.4.15-alpha.208) (2023-05-09)
+
+**Note:** Version bump only for package @hive-command/opcua-amqp-bridge
+
+
+
+
+
+## [1.4.15-alpha.207](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.206...v1.4.15-alpha.207) (2023-05-08)
+
+**Note:** Version bump only for package @hive-command/opcua-amqp-bridge
+
+
+
+
+
+## [1.4.15-alpha.206](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.205...v1.4.15-alpha.206) (2023-05-07)
+
+**Note:** Version bump only for package @hive-command/opcua-amqp-bridge
+
+
+
+
+
 ## [1.4.15-alpha.205](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.204...v1.4.15-alpha.205) (2023-05-07)
 
 **Note:** Version bump only for package @hive-command/opcua-amqp-bridge